requirements.txt
Django==5.2.5
dj-database-url==2.1.0
djangorestframework==3.14.0
django-filter==23.1
django-celery-beat==2.8.1
psycopg2-binary==2.9.6
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.15.0
prophet==1.1.4








